finalized the Navbar
