+++ src/component/Header/Logo/Logo.jsx
@@ -1,23 +1,17 @@
 import { Button, Typography } from '@material-ui/core';
 import MusicVideoIcon from '@material-ui/icons/MusicVideo';
 import React from 'react';
-import useStyles from './logo';
-// const useStyles = makeStyles({
-//   tool: {
-//     display: 'flex',
-//     justifyContent: 'space-around'
-//   }
-// });
+import HeaderStyles from '../HeaderStyles';
 
 const Logo = () => {
-  const classes = useStyles();
+  const classes = HeaderStyles();
 
   return (
     <div className={classes.tool}>
       <Button variant="contained" color="primary">
         <MusicVideoIcon />
       </Button>
-      <Typography variant="h6" color="secondary">
+      <Typography variant="h6" color="secondary" className={classes.logo}>
         S-Y-R
       </Typography>
     </div>
