--- src/component/Header/Logo/logo.js
@@ -1,10 +1,0 @@
-import { makeStyles } from '@material-ui/core/styles';
-
-const useStyles = makeStyles({
-  tool: {
-    display: 'flex',
-    justifyContent: 'space-around'
-  }
-});
-
-export default useStyles;
